requirements.txt
opencv-python
azure-storage-blob
Pillow
